+++ src/main/java/pl/pawc/ewi/model/RaportKilometry.java
@@ -4,13 +4,15 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
 public class RaportKilometry {
 
     private String maszynaid;
-    private double stanpoczatkowy;
+    private BigDecimal stanpoczatkowy;
     private String maszynanazwa;
 
 }
